pyproject.toml
[tool.poetry]
name = "python-langchain-examples"
version = "0.1.0"
description = ""
authors = [""]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.57.4"
langchain = "^0.3.12"
python-dotenv = "^1.0.1"
langchain-community = "^0.3.12"
chromadb = "^0.5.23"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
